requirements.txt
flask==3.0.0
flask-wtf==1.2.1
wtforms==3.1.1
pyyaml==6.0.1
openai==1.35.15
python-dotenv==1.0.0
anthropic==0.31.2
google-search-results==2.4.2
requests==2.31.0
beautifulsoup4==4.12.3
markdown==3.5.2
pdfkit==1.0.0 